:truck: renaming file modeDTO
