+++ bidding-strategy/src/biddingModule/loadPVservice.py
@@ -1,9 +1,9 @@
 import pandas as pd
 import numpy as np
-from biddingModule.mode_dto import Mode
+from biddingModule.modeDTO import Mode
 class StaticDataService:
     def __init__(self):
-        self.df=pd.read_csv('/home/bidding-strategy/data/data.csv')
+        self.df=pd.read_csv('/home/bidding-strategy/data/bldCharacteristicData.csv',parse_dates=True)
         self.num_round=len(self.df.index)
         pass
     
@@ -12,4 +12,8 @@ class StaticDataService:
             index = int(factor*self.num_round)
             train_df = self.df.loc[:index].reset_index()
             test_df = self.df.loc[index:].reset_index()
+            cols_of_interest = ['CHAM1-PV','CHAM2-PV','CHAM3-PV','CHAM4-PV','CHAM5-PV'] 
+            # print(len(test_df.index))
+            test_df=test_df[(test_df[cols_of_interest] != 0).any(axis=1)].reset_index()
+            # print(len(test_df.index))
         return train_df, test_df
