+++ bidding-strategy/src/biddingModule/modeDTO.py
